% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundance.R
\name{abundance}
\alias{abundance}
\title{Abundance}
\usage{
abundance(
  phylo_ob,
  level = "genus",
  id = "abcno",
  sample_id = "Time",
  relative_abun = TRUE,
  strata = NULL,
  strata_val = NULL,
  remove_collapsed_taxa = FALSE,
  select_taxa = NULL,
  select_type = NULL
)
}
\arguments{
\item{phylo_ob}{Phyloseq object with metadata in sample_data.}

\item{level}{Taxonomic rank from tax_table, case insensitive; default is "genus".}

\item{id}{Subject id; default is "abcno".}

\item{sample_id}{Sample identifier, to be used for repeated sampling of the same subject or different compartments; default is "Time".}

\item{relative_abun}{Use relative abundances, else absolute; default is TRUE.}

\item{strata}{Name of variable for stratification, e.g. "Time";}

\item{strata_val}{Value in variable @strata to keep; default is NULL.}

\item{remove_collapsed_taxa}{Will remove unclassified taxa which are elsewise merged; default is FALSE.}

\item{select_taxa}{Choose all taxa from a Taxonomic variable, eg. "Staphylocuccus" or "Staph" or "cuccus"; default is NULL.}

\item{select_type}{Taxonomic rank of the @select_taxa; default is NULL.}

\item{predictor}{Predictor of interestfor stastics/plotting in sample_data.}
}
\value{
A data.frame
}
\description{
For creating relative abundance data.frame
}
